Moved SpecialThreeDScene out of quaternions into three_d_scene
